# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CERBERUS Commerce OS API
  version: "4.1-production-readiness"
  description: >
    Çekirdek route-handler sözleşmesi. Uçlar varsayılan olarak HttpOnly
    `cerberus_session` JWT çereziyle korunur. Açık uçlar `/api/auth/login`,
    `/api/health` ve `/api/health/ready` olarak security alanında ayrıca işaretlenir.
    Kimliksiz/geçersiz oturum 401, rol veya mağaza kapsamı ihlali 403, body
    doğrulama hatası 422, aşırı gövde 413 ve beklenmeyen sunucu hatası
    correlationId ile 500 döner. STORE_USER için mağaza kapsamı sunucuda zorlanır.
servers:
  - url: /
security:
  - cookieAuth: []
...
paths:
...
  /api/orders/export:
    get:
      tags: [orders]
      summary: Filtrelenmiş siparişlerin tamamını CSV stream olarak indir
      description: >
        GET /api/orders ile aynı storeCode, cargoStatus, pshBatchNo ve search
        filtrelerini kullanır. STORE_USER kapsamı ve hassas alan maskelemesi export'ta
        da uygulanır. 250 satırlık parçalarla üretilir; 10.000 satır sessizce kesilmez.
      parameters:
        - { name: storeCode, in: query, schema: { type: string } }
        - { name: cargoStatus, in: query, schema: { type: string } }
        - { name: pshBatchNo, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string, maxLength: 120 } }
      responses:
        "200":
          description: UTF-8 BOM içeren 40 kolon CSV stream
          headers:
            X-Export-Row-Count:
              schema: { type: integer }
              description: Export edilen satır sayısı
          content:
            text/csv:
              schema: { type: string, format: binary }
        "401": { description: Oturum gerekli }
        "422": { description: Filtre sonucu 10.000 satırdan fazla; filtreyi daraltın }
...
  /api/orders/{id}:
    patch:
      tags: [orders]
      summary: Sipariş güncelle (beyaz liste alanlar; storeCode kapsam kontrolü)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: "message + order" }
        "401": { description: Oturum gerekli }
        "403": { description: Mağaza kapsamı dışı }
        "422": { description: Doğrulama hatası }
    delete:
      tags: [orders]
      summary: Sipariş sil (yalnız ADMIN/MANAGER)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: silindi }
        "401": { description: Oturum gerekli }
        "403": { description: Yetki yetersiz }
...
  /api/orders/import-drive-url:
    post:
      tags: [orders]
      summary: Google Drive XLS ayrıştırma (15 sn timeout, 20 MB üst sınır)
      responses:
        "200": { description: "headers + rows" }
        "401": { description: Oturum gerekli }
        "413": { description: Kaynak çok büyük }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cerberus_session
      description: HttpOnly, imzalı ve süreli JWT oturum çerezi
...